+++ packages/kit/src/exports/vite/index.js
@@ -34,8 +34,9 @@ import {
 	sveltekit_paths,
 	sveltekit_server
 } from './module_ids.js';
-import { resolve_peer_dependency } from '../../utils/import.js';
+import { import_peer } from '../../utils/import.js';
 import { compact } from '../../utils/array.js';
+import { crawlFrameworkPkgs } from 'vitefu';
 
 const cwd = process.cwd();
 
@@ -155,7 +156,7 @@ export async function sveltekit() {
 		...svelte_config.vitePlugin
 	};
 
-	const { svelte } = await resolve_peer_dependency('@sveltejs/vite-plugin-svelte');
+	const { svelte } = await import_peer('@sveltejs/vite-plugin-svelte');
 
 	return [...svelte(vite_plugin_svelte_options), ...(await kit({ svelte_config }))];
 }
@@ -182,7 +183,7 @@ let manifest_data;
  */
 async function kit({ svelte_config }) {
 	/** @type {import('vite')} */
-	const vite = await resolve_peer_dependency('vite');
+	const vite = await import_peer('vite');
 
 	const { kit } = svelte_config;
 	const out = `${kit.outDir}/output`;
@@ -228,7 +229,7 @@ async function kit({ svelte_config }) {
 		 * Build the SvelteKit-provided Vite config to be merged with the user's vite.config.js file.
 		 * @see https://vitejs.dev/guide/api-plugin.html#config
 		 */
-		config(config, config_env) {
+		async config(config, config_env) {
 			initial_config = config;
 			vite_config_env = config_env;
 			is_build = config_env.command === 'build';
@@ -250,6 +251,20 @@ async function kit({ svelte_config }) {
 			if (client_hooks) allow.add(path.dirname(client_hooks));
 
 			const generated = path.posix.join(kit.outDir, 'generated');
+
+			const packages_depending_on_svelte_kit = (
+				await crawlFrameworkPkgs({
+					root: cwd,
+					isBuild: is_build,
+					viteUserConfig: config,
+					isSemiFrameworkPkgByJson: (pkg_json) => {
+						return (
+							!!pkg_json.dependencies?.['@sveltejs/kit'] ||
+							!!pkg_json.peerDependencies?.['@sveltejs/kit']
+						);
+					}
+				})
+			).ssr.noExternal;
 
 			// dev and preview config can be shared
 			/** @type {import('vite').UserConfig} */
@@ -307,7 +322,11 @@ async function kit({ svelte_config }) {
 						//    when it is detected to keep our virtual modules working.
 						//    See https://github.com/sveltejs/kit/pull/9172
 						//    and https://vitest.dev/config/#deps-registernodeloader
-						'@sveltejs/kit'
+						'@sveltejs/kit',
+						// We need to bundle any packages depending on @sveltejs/kit so that
+						// everyone uses the same instances of classes such as `Redirect`
+						// which we use in `instanceof` checks
+						...packages_depending_on_svelte_kit
 					]
 				}
 			};
@@ -678,10 +697,10 @@ Tips:
 								sourcemapIgnoreList,
 								inlineDynamicImports: false
 							},
+							preserveEntrySignatures: 'strict',
 							onwarn(warning, handler) {
 								if (
 									warning.code === 'IMPORT_IS_UNDEFINED' &&
-									// TODO: rolldown doesn't seem to provide `warning.id` so this condition is never true https://github.com/rolldown/rolldown/issues/4427
 									warning.id === `${kit.outDir}/generated/client-optimized/app.js`
 								) {
 									// ignore e.g. undefined `handleError` hook when
@@ -713,16 +732,8 @@ Tips:
 					// }
 				};
 
-				if (!vite.rolldownVersion) {
-					// TODO: always set this once `preserveEntrySignatures` is implemented in rolldown
-					// https://github.com/rolldown/rolldown/issues/3500
-					// rolldown always does 'strict' in the meantime, so it's fine to only set this in non-rolldown mode
-					// @ts-ignore
-					new_config.build.rollupOptions.preserveEntrySignatures = 'strict';
-				}
-
 				if (!split && new_config.build?.rollupOptions?.output) {
-					const output_options = /** @type {import('vite').Rolldown.OutputOptions} */ (
+					const output_options = /** @type {import('vite').Rollup.OutputOptions} */ (
 						new_config.build.rollupOptions.output
 					);
 					if (vite.rolldownVersion) {
@@ -813,7 +824,7 @@ Tips:
 		 */
 		writeBundle: {
 			sequential: true,
-			async handler(_options) {
+			async handler(_options, bundle) {
 				if (secondary_build_started) return; // only run this once
 
 				const verbose = vite_config.logLevel === 'info';
@@ -844,26 +855,17 @@ Tips:
 					})};\n`
 				);
 
-				// first, build server nodes without the client manifest so we can analyse it
 				log.info('Analysing routes');
 
-				build_server_nodes(
-					out,
-					kit,
-					manifest_data,
-					server_manifest,
-					null,
-					null,
-					svelte_config.output
-				);
-
-				const metadata = await analyse({
+				const { metadata, static_exports } = await analyse({
 					hash: kit.router.type === 'hash',
 					manifest_path,
 					manifest_data,
 					server_manifest,
 					tracked_features,
-					env: { ...env.private, ...env.public }
+					env: { ...env.private, ...env.public },
+					out,
+					output_config: svelte_config.output
 				});
 
 				log.info('Building app');
@@ -878,7 +880,7 @@ Tips:
 
 				secondary_build_started = true;
 
-				const { output } = /** @type {import('vite').Rolldown.RolldownOutput} */ (
+				const { output: client_chunks } = /** @type {import('vite').Rollup.RollupOutput} */ (
 					await vite.build({
 						configFile: vite_config.configFile,
 						// CLI args
@@ -921,7 +923,7 @@ Tips:
 						imports: [...start.imports, ...app.imports],
 						stylesheets: [...start.stylesheets, ...app.stylesheets],
 						fonts: [...start.fonts, ...app.fonts],
-						uses_env_dynamic_public: output.some(
+						uses_env_dynamic_public: client_chunks.some(
 							(chunk) => chunk.type === 'chunk' && chunk.modules[env_dynamic_public]
 						)
 					};
@@ -970,14 +972,14 @@ Tips:
 						imports: start.imports,
 						stylesheets: start.stylesheets,
 						fonts: start.fonts,
-						uses_env_dynamic_public: output.some(
+						uses_env_dynamic_public: client_chunks.some(
 							(chunk) => chunk.type === 'chunk' && chunk.modules[env_dynamic_public]
 						)
 					};
 
 					if (svelte_config.kit.output.bundleStrategy === 'inline') {
-						const style = /** @type {import('vite').Rolldown.OutputAsset} */ (
-							output.find(
+						const style = /** @type {import('vite').Rollup.OutputAsset} */ (
+							client_chunks.find(
 								(chunk) =>
 									chunk.type === 'asset' &&
 									chunk.names.length === 1 &&
@@ -992,11 +994,6 @@ Tips:
 					}
 				}
 
-				const css = output.filter(
-					/** @type {(value: any) => value is import('vite').Rolldown.OutputAsset} */
-					(value) => value.type === 'asset' && value.fileName.endsWith('.css')
-				);
-
 				// regenerate manifest now that we have client entry...
 				fs.writeFileSync(
 					manifest_path,
@@ -1009,14 +1006,16 @@ Tips:
 				);
 
 				// regenerate nodes with the client manifest...
-				build_server_nodes(
+				await build_server_nodes(
 					out,
 					kit,
 					manifest_data,
 					server_manifest,
 					client_manifest,
-					css,
-					svelte_config.kit.output
+					bundle,
+					client_chunks,
+					svelte_config.kit.output,
+					static_exports
 				);
 
 				// ...and prerender
@@ -1054,7 +1053,9 @@ Tips:
 							...vite_config,
 							build: {
 								...vite_config.build,
-								minify: initial_config.build?.minify ?? vite.rolldownVersion ? 'oxc' : 'esbuild'
+								minify: initial_config.build?.minify ?? (
+									vite.rolldownVersion
+									? 'oxc' : 'esbuild')
 							}
 						},
 						manifest_data,